+++ molgenis-data-mapper/src/main/java/org/molgenis/data/mapper/repository/impl/MappingProjectRepositoryImpl.java
@@ -137,8 +137,12 @@ public class MappingProjectRepositoryImpl implements MappingProjectRepository
 	@Override
 	public void delete(String mappingProjectId)
 	{
-		List<MappingTarget> mappingTargets = getMappingProject(mappingProjectId).getMappingTargets();
-		dataService.delete(MappingProjectRepositoryImpl.META_DATA.getName(), mappingProjectId);
-		mappingTargetRepository.delete(mappingTargets);
+		MappingProject mappingProject = getMappingProject(mappingProjectId);
+		if (mappingProject != null)
+		{
+			List<MappingTarget> mappingTargets = mappingProject.getMappingTargets();
+			dataService.delete(MappingProjectRepositoryImpl.META_DATA.getName(), mappingProjectId);
+			mappingTargetRepository.delete(mappingTargets);
+		}
 	}
 }
